+++ includes/system/class-option.php
@@ -80,6 +80,7 @@ class Option {
 		self::$defaults['earlyloading']      = false;
 		self::$defaults['forceearlyloading'] = true;
 		self::$defaults['adminbar']          = true;
+		self::$defaults['maxsize']           = 100;
 		self::$network                       = [ 'earlyloading', 'version', 'use_cdn', 'use_apcu',  'script_in_footer', 'display_nag', 'analytics', 'history', 'gc', 'metrics', 'adminbar' ];
 	}
 	/**