+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [3.1.0] - Not Yet Released
 
 ### Added
+- New objects viewer in APCu management tools.
 
 ### Changed
 - Site Health page now presents a much more realistic test about object caching.